+++ src/components/FixedPlugin/FixedPlugin.jsx
@@ -1,5 +1,6 @@
 /*eslint-disable*/
 import React, { Component } from "react";
+import classnames from "classnames";
 
 import imagine1 from "assets/img/sidebar-1.jpg";
 import imagine2 from "assets/img/sidebar-2.jpg";
@@ -23,7 +24,7 @@ class FixedPlugin extends Component {
   }
   render() {
     return (
-      <div className="fixed-plugin">
+      <div className={classnames("fixed-plugin",{"rtl-fixed-plugin": this.props.rtlActive})}>
         <div id="fixedPluginClasses" className={this.props.fixedClasses}>
           <div onClick={this.handleClick}>
             <i className="fa fa-cog fa-2x" />