+++ src/components/Sidebar/Sidebar.jsx
@@ -48,15 +48,15 @@ const Sidebar = ({ ...props }) => {
           >
             <ListItem button className={classes.itemLink + listItemClasses}>
               {typeof prop.icon === "string" ? (
-                <Icon className={classes.itemIcon + whiteFontClasses}>
+                <Icon className={classNames(classes.itemIcon,whiteFontClasses,{[classes.itemIconRTL]: props.rtlActive})}>
                   {prop.icon}
                 </Icon>
               ) : (
-                <prop.icon className={classes.itemIcon + whiteFontClasses} />
+                <prop.icon className={classNames(classes.itemIcon,whiteFontClasses,{[classes.itemIconRTL]: props.rtlActive})} />
               )}
               <ListItemText
-                primary={prop.sidebarName}
-                className={classes.itemText + whiteFontClasses}
+                primary={props.rtlActive ? prop.sidebarRTLName:prop.sidebarName}
+                className={classNames(classes.itemText,whiteFontClasses,{[classes.itemTextRTL]: props.rtlActive})}
                 disableTypography={true}
               />
             </ListItem>
@@ -67,7 +67,7 @@ const Sidebar = ({ ...props }) => {
   );
   var brand = (
     <div className={classes.logo}>
-      <a href="https://www.creative-tim.com" className={classes.logoLink}>
+      <a href="https://www.creative-tim.com" className={classNames(classes.logoLink,{[classes.logoLinkRTL]: props.rtlActive})}>
         <div className={classes.logoImage}>
           <img src={logo} alt="logo" className={classes.img} />
         </div>
@@ -80,10 +80,10 @@ const Sidebar = ({ ...props }) => {
       <Hidden mdUp implementation="css">
         <Drawer
           variant="temporary"
-          anchor="right"
+          anchor={props.rtlActive ? "left":"right"}
           open={props.open}
           classes={{
-            paper: classes.drawerPaper
+            paper: classNames(classes.drawerPaper,{[classes.drawerPaperRTL]:props.rtlActive})
           }}
           onClose={props.handleDrawerToggle}
           ModalProps={{
@@ -105,11 +105,11 @@ const Sidebar = ({ ...props }) => {
       </Hidden>
       <Hidden smDown implementation="css">
         <Drawer
-          anchor="left"
+          anchor={props.rtlActive ? "right":"left"}
           variant="permanent"
           open
           classes={{
-            paper: classes.drawerPaper
+            paper: classNames(classes.drawerPaper,{[classes.drawerPaperRTL]:props.rtlActive})
           }}
         >
           {brand}