+++ src/index.js
@@ -4,8 +4,8 @@ import { createBrowserHistory } from "history";
 import { Router, Route, Switch, Redirect } from "react-router-dom";
 
 // core components
-import Admin from "layouts/Admin/Admin.jsx";
-import RTL from "layouts/RTL/RTL.jsx";
+import Admin from "layouts/Admin.jsx";
+import RTL from "layouts/RTL.jsx";
 
 import "assets/css/material-dashboard-react.css?v=1.5.0";