+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [1.5.0] 2018-09-21
+## [1.6.0] 2019-01-31
 ### Nice stuff
 ### Bug Fixing
 - [Added new script for npm](https://github.com/creativetimofficial/material-dashboard-react/issues/69)
@@ -6,6 +6,9 @@
 ### Major style changes
 - **src/assets/css/material-dashboard-react.css**(added styles for the new component, FixedPlugin)
 - **src/assets/jss/material-dashboard-react/components/customInputStyle.jsx**
+- **src/assets/jss/material-dashboard-react/layouts/rtlStyle.jsx**(new style file for the new layout)
+- **src/assets/jss/material-dashboard-react/views/rtlStyle.jsx**(new style file for the new page)
+- **src/assets/jss/material-dashboard-react/components/sidebarStyle.jsx**
 ### Major components changes
 - Changes caused by running [the prettier command](https://prettier.io/docs/en/install.html) for *.jsx*, *.js*, *.html* and *.css* files
 - **src/components/Sidebar/Sidebar.jsx**
@@ -14,9 +17,10 @@
 - **src/views/Notifications/Notifications.jsx**
 ### New components
 - **src/components/FixedPlugin/FixedPlugin.jsx**(Left menu used by us for customization - due to this, we've had to add a link for the fonts awesome cdn)
-- **src/layouts/RTL/RTL.jsx**
+- **src/layouts/RTL.jsx**
 - **src/views/RTLPage/RTLPage.jsx**
-- **src/layouts/Admin/Admin.jsx** (instead of Dashboard)
+- **src/layouts/Admin.jsx** (instead of Dashboard)
+- **src/components/Header/RTLHeaderLinks.jsx**
 ### Deleted components
 - **src/routes/dashboard.jsx**
 - **src/routes/index.jsx**